feat: sp-202412 회원 관리 기능 개선
회원 추가, 회원정보 수정

+++ admin/src/main/java/com/damso/admin/storage/repository/member/MemberRepository.java
@@ -3,6 +3,6 @@ package com.damso.admin.storage.repository.member;
 import com.damso.admin.storage.entity.member.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberRepository extends JpaRepository<Member, Integer> {
+public interface MemberRepository extends JpaRepository<Member, Long> {
     boolean existsByEmail(String email);
 }
